style: move `AsyncReplicationState` to `components`
